+++ common/src/main/java/com/pw/common/model/ActionType.java
@@ -1,0 +1,5 @@
+package com.pw.common.model;
+
+public enum ActionType {
+    MOVE, PICKUP, TEST, PLACE, DESTROY, SEND
+}
