+++ common/src/main/java/com/pw/common/model/TeamColor.java
@@ -1,0 +1,5 @@
+package com.pw.common.model;
+
+public enum TeamColor {
+    RED, BLUE
+}
