+++ common/src/main/java/com/pw/common/model/TeamRole.java
@@ -1,0 +1,5 @@
+package com.pw.common.model;
+
+public enum TeamRole {
+    LEADER, MEMBER
+}
