--- demo/src/main/java/com/pw/demo/DemoApp.java
@@ -1,14 +1,0 @@
-package com.pw.demo;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-
-@SpringBootApplication
-public class DemoApp {
-
-	public static void main(String[] args) {
-		new SpringApplicationBuilder()
-				.sources(DemoApp.class)
-				.run(args);
-	}
-}
