+++ game-master/src/main/java/com/pw/gamemaster/model/GameMasterBoard.java
@@ -1,12 +1,11 @@
 package com.pw.gamemaster.model;
 
-import com.pw.common.Board;
-import com.pw.common.Cell;
-import com.pw.common.Field;
-import com.pw.common.Position;
-import com.pw.player.model.Player;
-import com.pw.player.model.PlayerDto;
-import com.pw.player.model.Team;
+import com.pw.common.dto.PlayerDTO;
+import com.pw.common.model.Board;
+import com.pw.common.model.Cell;
+import com.pw.common.model.Field;
+import com.pw.common.model.Position;
+import com.pw.common.model.TeamColor;
 
 import java.awt.*;
 import java.util.HashSet;
@@ -17,24 +16,12 @@ import java.util.UUID;
 public class GameMasterBoard extends Board {
     private Set<Position> piecesPosition;
     public HashSet<UUID> piecesTested;
-    // since these values are to be inherited, im not sure if theyre needed here, someone confirm pls
-    //    public Cell[][] cellsGrid;
-    //    public int goalAreaHeight;
-    //    public int taskAreaHeight;
-    //    public int boardHeight = 2 * goalAreaHeight + taskAreaHeight;
-    //    public int boardWidth;
-
-    public enum PlacementResult {
-        CORRECT,
-        POINTLESS
-    }
 
     public GameMasterBoard() {
-
         super();
     }
 
-    public Position playerMove(PlayerDto playerDto, Position.Direction direction) {
+    public Position playerMove(PlayerDTO playerDTO, Position.Direction direction) {
         throw new RuntimeException("not implemented");
     }
 
@@ -52,17 +39,17 @@ public class GameMasterBoard extends Board {
 
     }
 
-    public PlacementResult placePiece(PlayerDto playerDto) {
+    public PlacementResult placePiece(PlayerDTO playerDTO) {
 
         throw new RuntimeException("not implemented");
     }
 
-    public Position placePlayer(PlayerDto playerDto) {
+    public Position placePlayer(PlayerDTO playerDTO) {
 
         throw new RuntimeException("not implemented");
     }
 
-    public boolean checkWinCondition(Team.TeamColor teamColor) {
+    public boolean checkWinCondition(TeamColor teamColor) {
 
         throw new RuntimeException("not implemented");
     }
