+++ game-master/src/main/java/com/pw/gamemaster/model/GameMasterStatus.java
@@ -1,0 +1,5 @@
+package com.pw.gamemaster.model;
+
+public enum GameMasterStatus {
+    ACTIVE, FINISHED, IDLE
+}
