+++ game-master/src/main/java/com/pw/gamemaster/model/PlacementResult.java
@@ -1,0 +1,5 @@
+package com.pw.gamemaster.model;
+
+public enum PlacementResult {
+    CORRECT, POINTLESS
+}
