+++ player/src/main/java/com/pw/player/model/PlayerState.java
@@ -1,0 +1,5 @@
+package com.pw.player.model;
+
+public enum PlayerState {
+    INITIALIZING, ACTIVE, COMPLETED
+}
