+++ player/src/main/java/com/pw/player/model/Team.java
@@ -1,32 +1,18 @@
 package com.pw.player.model;
 
+import com.pw.common.model.TeamColor;
+import com.pw.common.model.TeamRole;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
+@AllArgsConstructor
 public class Team {
     public TeamColor color;
     public TeamRole role;
     public int size;
 
-
-    public enum TeamRole {
-        LEADER,
-        MEMBER
-    }
-
-    public enum TeamColor {
-        RED,
-        BLUE
-    }
-
-    public Team(TeamRole teamRole, TeamColor teamColor, int size) {
-        this.color = teamColor;
-        this.size = size;
-        this.role = teamRole;
-    }
-
     public boolean isLeader() {
-
         return role == TeamRole.LEADER;
     }
 }
