+++ pom.xml
@@ -83,7 +83,7 @@
 		<module>game-master</module>
 		<module>player</module>
 		<module>common</module>
-		<module>demo</module>
+		<module>server</module>
 	</modules>
 
 </project>
