+++ server/src/main/java/com/pw/server/ServerApp.java
@@ -1,0 +1,8 @@
+package com.pw.server;
+
+public class ServerApp {
+
+	public static void main(String[] args) {
+		System.out.println("dupa");
+	}
+}
